Refactor App component: extract ProductList, Cart, and Checkout into separate components for improved readability and maintainability
